Cleaned up branch 1.2
Cleaned up any file that are not needed, I'm working at the trunk of this branch.

Potentials, distributions and tables seem to work perfectly. Added all the necessary test cases.

I use the new definition for multiplication: in-place and we add any new dimensions to a.

Newt thing to do: check why JunctionTree results are not correct...