Youhou!!!!
JTree is now working perfectly.
Results are correct for the Water-Sprinkler (with/without evidence) and the algorithm seems to work for a more general bayesian structure.

Should add a couple of test case with a more complex bayesian network to be sure.

Most of the errors that I corrected were due to using names (which is a set, with no order predefined) instead of names_list (with has the correct order). I think I fixed all of them.

I found out that when evidence is enetered, although the results are correct more messages than needed are sent through th network. Should chekc that out.