Corrected initialisation error of the Junction tree Inference Engine.
If a node of the bayesian network is present in more than one clusters of the junction tree, then it is counted more than once at the initialization part of the algorithm resulting in erroneous calculations. The bug was easily fixed by adding a 'break' in the initialization loop, once a variable has already been treated. 

This bug was reported by Michael Munie munie@stanford.edu. 
Thank you for your contribution Michael