Corrected the bug  reported by Ivan. The error was inside the Triangulasiation algorithm of the Junction Tree, some important edges where not added to the graph resulting in missing dependancies between the variables.
Also, cleared the code a little bit and corrected some mistakes in the TestCases. Now all the TestCases should work. However, in the distributions.py, some tests fail once every 10 times because the order of the variables is not set. Running the test again should produce the correct results. This will be fixed in a later time.

Created a new copy function that is delegated throughout the graph --> BNet classes